+++ dnssec-tools/validator/apps/validator_driver.h
@@ -7,11 +7,11 @@
 
 void print_val_response(struct val_response *resp);
 int sendquery(val_context_t * context, const char *desc, u_char * name_n,
-              const u_int16_t class, const u_int16_t type, u_int8_t flags,
+              const u_int16_t class, const u_int16_t type, u_int32_t flags,
               const int *result_ar, int trusted_only,
               struct val_response **resp);
 
-int self_test(val_context_t *context, int tcs, int tce, u_int8_t flags,
+int self_test(val_context_t *context, int tcs, int tce, u_int32_t flags,
               const char *tests, const char *suites, int doprint);
